# contracts/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Jobseeker Copilot - Document Store API",
    "description" : "Service for storing and retrieving generated CVs and cover letters.\n\nThis service stores generated document content.\nIt does NOT generate documents.\nIt does NOT call an LLM.\nIt does NOT export PDF/DOCX.\n\nEvery document and file operation is authenticated and owner scoped.\nUser ownership comes from a validated access-token subject. Approved\nservice identities must bind the same owner in X-Document-Owner.\n\nDocumentType values:\n- CV - Curriculum Vitae / Resume\n- COVER_LETTER - Cover letter for a job application\n",
    "contact" : {
      "name" : "Jobseeker Copilot"
    },
    "license" : {
      "name" : "Proprietary and confidential"
    },
    "version" : "1.2.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8089",
    "description" : "Local development"
  } ],
...
  "paths" : {
    "/internal/system-data/seed/documents" : {
      "post" : {
        "tags" : [ "document-store-system-data-controller" ],
        "operationId" : "seedDocuments",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SystemDataDocumentSeedRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SystemDataResult"
                }
              }
            }
          }
        },
        "security" : [ {
          "environmentDataToken" : [ ]
        } ]
      }
    },
...
    "/internal/system-data/verify/documents/{userId}" : {
      "get" : {
        "tags" : [ "document-store-system-data-controller" ],
        "operationId" : "verifyDocuments",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SystemDataResult"
                }
              }
            }
          }
        },
        "security" : [ {
          "environmentDataToken" : [ ]
        } ]
      }
    },
...
    "/internal/system-data/scenario/{scenarioId}/documents/{userId}" : {
      "delete" : {
        "tags" : [ "document-store-system-data-controller" ],
        "operationId" : "resetDocuments",
        "parameters" : [ {
          "name" : "scenarioId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SystemDataResult"
                }
              }
            }
          }
        },
        "security" : [ {
          "environmentDataToken" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "GeneratedDocument" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string"
          },
          "jobId" : {
            "type" : "string"
          },
          "applicationId" : {
            "type" : "string"
          },
          "documentType" : {
            "type" : "string",
            "enum" : [ "CV", "COVER_LETTER" ]
          },
          "title" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "active" : {
            "type" : "boolean"
          },
          "originalFilename" : {
            "type" : "string"
          },
          "sourceType" : {
            "type" : "string",
            "enum" : [ "GENERATED", "UPLOADED" ]
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "SystemDataDocumentFileSeed" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "generatedDocumentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "fileType" : {
            "type" : "string",
            "enum" : [ "DOCX", "PDF" ]
          },
          "fileName" : {
            "type" : "string"
          },
          "mimeType" : {
            "type" : "string"
          },
          "source" : {
            "type" : "string",
            "enum" : [ "GENERATED", "USER_UPLOADED" ]
          },
          "active" : {
            "type" : "boolean"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "fileContent" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "byte"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "SystemDataDocumentSeedRequest" : {
        "type" : "object",
        "properties" : {
          "scenarioId" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          },
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GeneratedDocument"
            }
          },
          "files" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SystemDataDocumentFileSeed"
            }
          }
        }
      },
      "SystemDataResult" : {
        "type" : "object",
        "properties" : {
          "service" : {
            "type" : "string"
          },
          "operation" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "recordsAffected" : {
            "type" : "integer",
            "format" : "int32"
          },
          "activeEnvironment" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "warnings" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "serviceToken" : {
        "type" : "apiKey",
        "name" : "X-Service-Token",
        "in" : "header"
      },
      "environmentDataToken" : {
        "type" : "apiKey",
        "name" : "X-Environment-Data-Token",
        "in" : "header"
      }
    }
  }
}